Fix bug #133; Logo no longer crashes when marcos output nothing.
I chose to fix this by reducing the times in which the DIDNT_OUTPUT error is thrown from eval_sequence_continue, instead of removing it completely, as is done in UCBLogo, because removing it introduce new bugs and patching those would have required porting too much code from UCBLogo.  Further testing may show this fix to be inadequate, in which case I'll port more of the fix from UCBLogo to FMSLogo.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3536 9d0050d7-00b0-40f9-871c-8e02cbd5794f