Move the logic for the event handlers for all of the context menu items which CLogoCodeCtrl implements from CWorkspaceEditor to CLogoCodeCtrl.  This was done to make CLogoCodeCtrl self-sufficient so that it could be more easily used in other places.
Also, change the signature of SetFont to override the virtual function inherited from the base class.  This was done so thata caller wouldn't accidently call the wrong one (which would have no effect).

Both of these changes were made in anticipation of making the commander's input a CLogoCodeCtrl, instead of a wxTextCtrl.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3601 9d0050d7-00b0-40f9-871c-8e02cbd5794f