First pass at fixing bug #347; copy the image bitmap and the bitmapped turtles onto a separate bitmap before blitting it to the screen.  This fixes a problem where the bitmaps would flicker if the vsync happened between when the image was copied and when the bitmap was copied over it.
The back buffer is used only when there is at least one rotating bitmap turtle so that programmers who are not using this feature don't pay a performance cost.

Because copying to a back buffer is done 1:1 with the source bitmap size and then blitted using a stretch blit (as opposed to before, where the rotating bitmap was rotated and scaled directly onto the screen), the resulting code is not any slower than before.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3751 9d0050d7-00b0-40f9-871c-8e02cbd5794f