Port Asteroid Miner to use rotating bitmaps for the ship.  This was partly done  to show an example of the new feature, partly to prove that it is complete enough to be used in a game, and partly to make Asteroid Miner run fast enough to be playable on my 90 Mhz Pentium.
Overall, this was not the success I was hoping for.  It runs faster, but not fast enough.  It didn't make the code any simpler, since the bitmaps are still drawn within Logo.  And it caused a regression in some features that were not easy to do within the sprite framework, such a gradual fade effect when picking up a gem and a shrinking effect and the shrinking ship effect when beating a level.  For gameplay, preserving the exact effect is not that important, but having some kind of effect is.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3764 9d0050d7-00b0-40f9-871c-8e02cbd5794f