Fix bug #423; Menu items in the editor of the wxWidgets version of FMSLogo are now disabled when they would have no effect.
The problem was due to a bug in wxWidgets where UPDATE_UI events wouldn't be sent if it were compiled without status bar support, which I had configured out of my static library to reduce the size of the final binary.  I now build wxWidgets with support for status bars.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3861 9d0050d7-00b0-40f9-871c-8e02cbd5794f