Fix bug #424; Captial letters and other printable characters are now redirected from the screen to the commander in the wxWidgets version of FMSLogo.
I couldn't find a clean way to emulate a wxKeyEvent on a Scintilla control, as the wxStyledText wrapper hides the raw HWND and has no API for this.  The best  I could come up with is to set focus on the input and then generate a key event using a win16 API when sends the key event to Windows, not a particular window.

This change undoes a previous fix for a problem that was reported on Vista, but from the looks of the comment, this fix predates using a wxStyledTextCtrl, as it refers to an API that's available on wxTextCtrl, but not wxStyledTextCtrl.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3883 9d0050d7-00b0-40f9-871c-8e02cbd5794f