Implement feature #106; Error messages are now rendered in red text in the commander history box.
Each message that is logged to the commander history can now be tagged by the caller with a type and each FMSLogo front-end can do whatever it wants with this additional information.  All versions render the text is red and the OWL version also makes it bold.

This required adding a flag to many different printing functions and many control paths won't end up using it (for example, those that end up being directed to a file).


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3889 9d0050d7-00b0-40f9-871c-8e02cbd5794f