Change ARITY to throw a special "don't like" error for TO or .MACRO, instead of outputting [-1 -1 -1].
This also updates all untranslated strings to be whatever they are currently in English.  This way, if a string hasn't been translated, the translater will see the most recent text (which is what I'd like the to translate) instead of an older version that I have since decided wasn't good enough.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3914 9d0050d7-00b0-40f9-871c-8e02cbd5794f