Update the screensaver's Makefile to use the mingw-w64, instead of the older mingw.org compiler.  This is being done to standardize on a single compiler, which will hopefully make building FMSLogo simple enough that other people will try.
Unfortunately, the mingw-w64 suite has an empty libscrnsave.a, so upgrading the compiler required adding source code that should have been included in the library.  The scrnsave.c file is copied directly from the mingw-w64 source tree with an "#if 0" removed.  If mingw-w64 eventually has a functional libscrnsave.lib, this file should be removed.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4000 9d0050d7-00b0-40f9-871c-8e02cbd5794f