Fix a bug in the WM_ERASEBKGND message handler of the screensaver that caused it to erase the entire screen, instead of the region that needed erasing.  My fix is redundant, in that merely using the HDC that was passed in as the WPARAM appeared to be sufficient (as the HDC presumably is clipping the parts that don't need updated).  However to be sure, the screensaver only erases the rectangle that is invalid.
While this bug has existed since the initial release of the FMSLogo screensaver, it only became noticable now, as I recently implemented the windowing command set in the screensaver, which makes it possible to show/hide windows on top of the screensaver.  When a window is hidden, the WM_ERASEBACKGROUND message is sent to repaint the region where the window was.  Due to the bug, the entire screen would be made white, then only the region under the window would be repainted with WM_PAINT.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4024 9d0050d7-00b0-40f9-871c-8e02cbd5794f