Remove references to the GiveFocusToEditbox global variable.
Once upon a time, this was used to give focus to the commander's edit box after the "Edall" button was pressed.  This was a vestige from MSWLogo, which was compiled using Borland OWL and was unable to redirect keystrokes from the Edall button to the commander input after the Edall button had focus.  In the current wxWidgets implementation, this isn't needed and its existence made the code unnecessarily more difficult to understand.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4056 9d0050d7-00b0-40f9-871c-8e02cbd5794f