Fix bug #472; SETPIXEL now behave correctly, even in PENREVERSE mode
This was a regression that was introduced in FMSLogo 6.16.0 when I improved drawing performance by using the same screen DC and memory DC for all operating (instead of recreating them for each operation).  As a result of this change, the device contexts retained their state between calls, including their ROP2 mode.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4065 9d0050d7-00b0-40f9-871c-8e02cbd5794f