Update FMSLogo from wxWidgets 2.8.12 to 3.1.0
While there's nothing in the new library version that is immediately needed, it's good to stay on the latest version.  Since I'm incrementing the major version number of FMSlogo, it's a good time to include new libraries which may introduce incompatibilities.

This has an unfortunately side-effect of doubling the size of the executable, from 5MB to 10MB.  This can likely be reduced by figuring what new features of 3.1.0 are being included that are not needed.

This requires some minor changes to the way the FMSLogo lexer is registered with Scintilla.  This used to be automatic, as LexerModule was an inline linked list and creating a new one added it to the list.  Now this requires explicitly adding the LexerModule to a catalog.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4067 9d0050d7-00b0-40f9-871c-8e02cbd5794f