Enable wxStatusBar so that menu items can be disabled.
This is a regression introduced after rewriting the configure script to start by enabling features used instead of disabling ones that weren't used.  Since FMSLogo does not use wxStatusBar, it made sense not to enable it.  However, as described in bug #423, there's a quirk of wxWidgets where it requires the status bar to send "update UI events" that are used to disable menu items.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4079 9d0050d7-00b0-40f9-871c-8e02cbd5794f