Remove code in NOASM and NOASMX and remove the precompiler option SHARED.  This was done to simplify the code for future maintainers.
The NOASM and NOASMX are relics from when the code was compiled with Borland C++ and were not used to because they disabled assembly, but rather because it was debug code and the debug build also defined NOASM and NOASMX.  (The "X" suffix convention seems to be used to create a second symbol when the one is already in use).  If the someone needs to debug the code, it's unlikely that they will benefit from what George Mills used to debug the code and they can add whatever they need.

The SHARED code is what's used to improve performance (when splitting polygons with a plane, you end up with two identical vertices, so there's no need to rotate them separately).  The SHARED code is always compiled and it won't be removed, so having a #define SHARED only makes the code more difficult to read.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4101 9d0050d7-00b0-40f9-871c-8e02cbd5794f