Fix a problem with SELECTBOX where it didn't appear when called from the screensaver (actually it did appear, but it was behind the screen window, so it couldn't be seen).
Adding this to the screesaver test required refactoring the SELECTBOX tests to split the manual tests from the fully automated ones, as manual tests require the use of a mouse and moving the mouse exits the screensaver.  Because of this, SELECTBOX is not likely to be useful in a screensaver, but this is a decision that a programmer can make.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4137 9d0050d7-00b0-40f9-871c-8e02cbd5794f