Fix a regression in the screensaver where using DIALOGFILEOPEN created the window behind the screen window, making it invisible.  The regression was introduced when the screensaver was ported to wxWidgets.
I had been fixing this class of bug by using the wxSTAY_ON_TOP style for each of the dialog boxes that a Logo command can create, but ran into a problem in DIALOGFILEOPEN where this had no effect (the dialog isn't a real wxWidgets dialog box, but a proxy to a common control).  After some experimentation, I settled on a new strategy, which is to lower the screensaver window so that any windows created after it would be on top of it.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4140 9d0050d7-00b0-40f9-871c-8e02cbd5794f