Remove some unneeded #defines that duplicated what should have been in windows.h, but now generated compiler warnings.
This was needed to be able to compile the screensaver with an old version of MinGW whose libc didn't include instructions that weren't implemented in the original Pentium CPUs.  As of version 7.0, FMSLogo no longer supports those machines or that compiler, so #defines can be safely removed (and the compiler warnings along with them).


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4145 9d0050d7-00b0-40f9-871c-8e02cbd5794f