Fix bug #484; BYE now immediately exits when run at the top-level of a file that is loaded from the command-line.
The heart of the fix is to make all primitives return an Unbound value when exiting to avoid the "you don't say what to do with" errors.  The rest of the changes udpate all parts that of the code that evaluate Logo code to do so within a start_execution() and stop_execution() so that a "close" event can be sent if the application is exiting.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4249 9d0050d7-00b0-40f9-871c-8e02cbd5794f