Fix bug #486; The turtle can now walk along the screen extents in FENCE mode.
The problem was that if you tried to walk along the extents, due to the impercision of floating point arithemetic, you might end up crossing the extent by 1e-14 or something similarly small (especially after conversion from degress to radians).

My fix was to permit the same "rounding" tolerance that POS uses when showing the true turtle position.  My thinking is that this tolerance is used to mask very small amounts of user-facing of imprecision that are the turtle's position, the screen extent is another user-facing boundary, so it should use the same precision.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4264 9d0050d7-00b0-40f9-871c-8e02cbd5794f