Fix Bug #489; Running a Logolib procedure onces again works when the current working directory is on a different drive than the FMSLogo executable.
The problem was that the base directory was extracted without the drive letter, so it was treated as relative to the current working directory's drive.  If the user changed the working directory to a different drive, then the logolib procedure would be loaded with the correct path on the wrong drive and fail.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4321 9d0050d7-00b0-40f9-871c-8e02cbd5794f