Replace the wxStyledTextCtrl implementation of CLogoCodeCtrl with wxTextCtrl for languages that use multi-byte character sets.  Neither wxStyledTextCtrl nor wxRichTextCtrl supported these character sets.
This is a step backward in overall functionality (no styling, no printing, no searching, no paren matching), but fixes the basic blocking problem of not being able to use characters from Simplified Chinese in their programs.  Some additional enhancements may be made to it, but not many, since that time would be better spent porting FMSLogo to Unicode.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4327 9d0050d7-00b0-40f9-871c-8e02cbd5794f