Update the expected test results based on the FMSLogo 7.1.0 test pass.
In this case, I modified the results from the DLLCALL buffer test because it put NUL characters into the output file, which caused my diffing tool to think that it was a binary file and so refused to compare it with the baseline.   The test will be fixed not to do this before the next test pass.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4362 9d0050d7-00b0-40f9-871c-8e02cbd5794f