Fix a case problem in the color name "DarkOrange".
This would have caused problems when setting a color as "Darkorange
and :ignorecase was defined to "TRUE (that is, it probably didn't make
a difference).  In that sense, this is a breaking change, but the likelihood
of a program depending on "DarkOrange" being called "Darkorange" is so low
that it's worth fixing.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4375 9d0050d7-00b0-40f9-871c-8e02cbd5794f