Switch from blocking undocumented procedures to only allowing documented procedures in the translation tables.
While this makes the code cleaner and more correct, it was primarily motivated by the fix for Bug #493, which put new COPYDEF calls into startup-1040.logoscript that were part of the implementation of SLOWDRAW, and not for translation purposes.

This also fixed a few mistakes in the Portugese and German translation tables, but since these use a custom manual, the translation tables were ignored.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4382 9d0050d7-00b0-40f9-871c-8e02cbd5794f