Switch to the FMSLogo base directory before loading startup.logoscript when erasing the workspace.
Fix a slight regression caused by the fix for Bug #492, which is that a COPYDEF for a library procedure would have the side-effect of trying to load a logo file with the given procedure's name.  This is probably unlikely to happen in practice, but during testing, it's very noticable since the "test" directory is full of files that are named after the library procedure which they test.  So if the current working directory was "test" and a typical translation startup.logoscript was loaded, File -> New would end up populating the workspace with a bunch of test procedures, instead of erasing them.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4405 9d0050d7-00b0-40f9-871c-8e02cbd5794f