Clarify what was meant by "The start position is always on the same ellipse, regardless of startangle".
When I read this recently, I couldn't understand what it meant or how the following sentence was explaining it.  Instead of combining to unrelated ideas ("start position" and "same ellipse"), I removed the "same ellipse" explanation from the startangle description and added it as a conclusion to the inline/crosswise paragraph.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4421 9d0050d7-00b0-40f9-871c-8e02cbd5794f