Fix Bug #499; When a single widget is added to a dialog or window from the setup instructionlist, it no longer fills the entire dialog/window.
This was a regression introduced in FMSLogo 7.0.0 caused by upgrading wxWidgets from 2.8.12 to 3.1.0.  In this upgrade, there was a bug fix in wxWidgets that made it behave this way (that it, it was always supposed to do so, but it so happened than in 2.8.12 it didn't due to a bug).

The fix for the FMSLogo bug is to add a sizer to all wxDialog that are created with WINDOWCREATE or DIALOGCREATE.  This disable the "single widget fills all space" behavior in wxWidgets.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4425 9d0050d7-00b0-40f9-871c-8e02cbd5794f