Translate the "Main Menu" section of the manual to Spanish. This is primarily based on the translation for MSWLogo by Rosario Ruiz.
While documenting some of the UI items and commands, it made sense to
fix the existing translations in the executable than to document them
as-is.  In particular "colorf" was translated as "colorFondo" for
symmetry with "ponColorFondo" and "pintaColor" was changed to "colorRelleno"
for symmetry with the commands "ponColorRelleno" and "rellena".


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4479 9d0050d7-00b0-40f9-871c-8e02cbd5794f