Fix bug #515; PORTWRITEARRAY now throws a "doesn't like input" error for its second input.
The motivation for this change is that PORTWRITEARRAY used to show a message which incorrectly referred to the first input when given a non-array second input.  I could have instead corrected the text, but that would have left the non-English translations incorrect.  This fix makes PORTWRITEARRAY act more like a typical FMSLogo procedure, throwing an error instead of displaying a dialog box that cannot be interacted with programmatically.

Since, after showing the dialog box with the incorrect message, PORTWRITEARRAY used to throw a "Stopping..." error, changing this to throwing a "Doesn't like input" error has negligable risk to backward compatibility.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4564 9d0050d7-00b0-40f9-871c-8e02cbd5794f