Remove a repeated definition of BNOMBRES. From the copydef comment, it looks like this was supposed to be named BOVARS. This would have made more sense given the names BOVAR and LVARS.
Since BNOMBRES exist and BOVARS hasn't existed for a decade without complaint,
I don't think adding a BOVARS would improve FMSLogo, I'm simply deleting the
redundant definition and the errant copydef.

The impact on FMSLogo should only be the removal of a bogus entry
in the Spanish manual's translation tables.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4654 9d0050d7-00b0-40f9-871c-8e02cbd5794f